refactor(l5): move rest-promotion from src/ into server/transport (issue 864)
Moves the promotion REST router out of src/ and under server/, per L5 of
_plans/server-hardening-ladder.md.

Moved path:
- src/rest-promotion.ts -> server/transport/rest-promotion.ts

Runtime closure of src/ under server/main.ts: 24 -> 23.

The runtime importer server/transport/rest.ts:35 is repointed at the new
sibling path. Every other importer (src/index.ts:20, src/rest-promotion.test.ts,
src/ob-admin-role.test.ts) keeps its line: the old src/ path is now a shim
re-exporting the server/ module, so those files stay unstaged.

Dependency imports resolve to the server/ copy where the src/ path is already
an L5 shim: read-policy, namespace-policy, promotion-nomination, and the
observability log surface.

Three dependencies deliberately keep their src/ paths, because the server/
versions are not shims but distinct implementations that require the
composition root to inject config, and this route's caller
(server/transport/rest.ts) does not carry config:

- src/promotion-service.ts is itself an L5 adapter and reads
  OPENBRAIN_PROMOTION_KILL_SWITCH at call time;
  server/domain/promotion-service.ts takes killSwitch as a value. Calling the
  service directly dropped the kill switch and turned "blocks mutating
  promotion when the kill switch is enabled" red.
- src/shared-namespace.ts resolves its own names;
  server/tools/shared-namespace.ts throws "shared-namespace names are missing"
  unless the composition root supplies sharedNamespaceNames, which failed every
  scan and promote with a non-JSON 500.
- src/types.ts is an independent module, not a shim.

Repointing those three is L6 work, once config reaches the REST surface. Each
is recorded in a comment at its import.

The moved file is brought to the server/ lint standard by mechanical hoisting
only, with no behavior change: the three route handlers become module-scope
functions, the shared promoter-role guard and the promote failure reporter are
hoisted out of the duplicated bodies, the scan pass is split into scanTable and
findTargetDuplicate with the source window and target namespace grouped into
named types, and the two `any` uses become declared types. Query order, SQL,
status codes, and every error string are unchanged.

Tests stay at their src/ paths untouched per the M1 exception and exercise the
moved code through the shim.

Verified:
- bunx tsc --noEmit -> exit 0
- oxlint --deny-warnings on all three staged paths -> exit 0
- bun run test:isolated src/rest-promotion.test.ts src/ob-admin-role.test.ts
  -> 40 pass, 0 fail, matching the origin/main baseline
- bash scripts/done-means/864-moved-out-of-src.sh -> PASS
